+++ platform/dep/sync/service.go
@@ -2,6 +2,7 @@ package sync
 
 import (
 	"context"
+	"time"
 )
 
 type Service interface {
@@ -20,4 +21,20 @@ type DB interface {
 type DEPSyncService struct {
 	db     DB
 	syncer Syncer
+}
+
+type Cursor struct {
+	Value     string    `json:"value"`
+	CreatedAt time.Time `json:"created_at"`
+}
+
+// A cursor is valid for a week.
+func (c Cursor) Valid() bool {
+	expiration := time.Now().Add(cursorValidDuration)
+	return c.CreatedAt.Before(expiration)
+}
+
+type AutoAssigner struct {
+	Filter      string `json:"filter"`
+	ProfileUUID string `json:"profile_uuid"`
 }
